+++ client/jetpack-connect/persistence-utils.js
