+++ client/state/jetpack-connect/actions.js
@@ -51,7 +51,7 @@ import addQueryArgs from 'lib/route/add-query-args';
 import { externalRedirect } from 'lib/route/path';
 import { urlToSlug } from 'lib/url';
 import { JPC_PLANS_PAGE } from './constants';
-import { clearPlan } from 'jetpack-connect/cookies';
+import { clearPlan } from 'jetpack-connect/persistence-utils';
 
 /**
  *  Local variables;
